LPD-76454 Reorganize folder structure
