+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/FinanceDashboard/FinanceDashboardRouter.tsx
@@ -7,10 +7,10 @@ import {Route, Routes} from 'react-router-dom';
 
 import withProviders from '../../hoc/withProviders';
 import FinanceDashboardOutlet from './FinanceDashboardOutlet';
-import OrderDetails from './pages/Orders/OrderDetails/OrderDetails';
-import Orders from './pages/Orders/Orders';
-import PaymentDetails from './pages/Payments/PaymentDetails/PaymentDetails';
-import Payments from './pages/Payments/Payments';
+import Orders from './pages/Orders';
+import OrderDetails from './pages/Orders/OrderDetails';
+import Payments from './pages/Payments';
+import PaymentDetails from './pages/Payments/PaymentDetails';
 
 import './index.scss';
