+++ workspaces/liferay-marketplace-workspace/client-extensions/liferay-marketplace-custom-element/src/pages/FinanceDashboard/pages/Payments/PaymentDetails.tsx
@@ -6,21 +6,21 @@
 import ClayTable from '@clayui/table';
 import {useParams} from 'react-router-dom';
 
-import {DetailedCard} from '../../../../../components/DetailedCard/DetailedCard';
-import {PageRenderer} from '../../../../../components/Page';
-import QATable, {Orientation} from '../../../../../components/QATable';
-import Table from '../../../../../components/Table/Table';
-import {PaymentStatus as PaymentStatusCode} from '../../../../../enums/Order';
-import usePublisherSalesSummaryObject from '../../../../../hooks/usePublisherSalesSummaryObject';
-import i18n from '../../../../../i18n';
-import {Liferay} from '../../../../../liferay/liferay';
-import PublisherSalesSummary from '../../../../../services/rest/PublisherSalesSummary';
-import {safeJSONParse} from '../../../../../utils/util';
-import DetailsHeader from '../../../components/DetailsHeader/DetailsHeader';
-import PaymentStatus from '../../../components/PaymentStatus/PaymentStatus';
-import {formatCurrency, getTotalByOrderKey} from '../../../util/finance';
-import {formatDate, textWrapper} from '../../../util/util';
-import {PublisherPayoutStatus} from '../Payments';
+import {DetailedCard} from '../../../../components/DetailedCard/DetailedCard';
+import {PageRenderer} from '../../../../components/Page';
+import QATable, {Orientation} from '../../../../components/QATable';
+import Table from '../../../../components/Table/Table';
+import {PaymentStatus as PaymentStatusCode} from '../../../../enums/Order';
+import usePublisherSalesSummaryObject from '../../../../hooks/usePublisherSalesSummaryObject';
+import i18n from '../../../../i18n';
+import {Liferay} from '../../../../liferay/liferay';
+import PublisherSalesSummary from '../../../../services/rest/PublisherSalesSummary';
+import {safeJSONParse} from '../../../../utils/util';
+import DetailsHeader from '../../components/DetailsHeader/DetailsHeader';
+import PaymentStatus from '../../components/PaymentStatus/PaymentStatus';
+import {formatCurrency, getTotalByOrderKey} from '../../util/finance';
+import {formatDate, textWrapper} from '../../util/util';
+import {PublisherPayoutStatus} from '.';
 
 enum Payouts {
 	MP_COMMISSION = 0.2,
